Fix Insecure Direct Object Reference in parser error report download (pentest findings) (#2841)
* move xlsx serialization to datafiles serializers

* create new parser error download endpoint under datafiles viewset

* move xlsx serializer to utils

* move tests to test_api

* rm unneeded code

* rm comments

* rm blank lines

* fix test

* fix lint

* lint

* removed unused datafilesummary viewset

* missed a ref

---------
