+++ tdrs-backend/tdpservice/urls.py
@@ -38,7 +38,6 @@ urlpatterns = [
     path("stts/", include("tdpservice.stts.urls")),
     path("data_files/", include("tdpservice.data_files.urls")),
     path("logs/", write_logs),
-    path("parsing/", include("tdpservice.parsers.urls")),
 ]
 
 if settings.DEBUG:
